Setup separate rules and configs diretories
