+++ lib/configs/index.js
@@ -1,0 +1,7 @@
+const recommended = require("./recommended");
+const cypress = require("./cypress");
+
+module.exports = {
+  recommended,
+  cypress,
+};
