+++ lib/rules/index.js
@@ -1,4 +1,4 @@
-const noSingleLetterVariable = require("./rules/no-single-letter-variable");
+const noSingleLetterVariable = require("./no-single-letter-variable");
 
 module.exports = {
   "no-single-letter-variable": noSingleLetterVariable,
